requirements.txt
selenium==4.9.0
pillow==9.5.0
pyobjc>=10.3
pyobjc-core>=10.3
numpy
pyautogui
colorama
openai
#sorry for using this non mac people
ocrmac
thefuzz

#this likely isnt valid